// src/api/openapi/models/UserFinancialSettingsResponseDto.ts
export interface UserFinancialSettingsResponseDto {
    /**
     * 
     * @type {string}
     * @memberof UserFinancialSettingsResponseDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof UserFinancialSettingsResponseDto
     */
    surname: string;
    /**
     * 
     * @type {string}
     * @memberof UserFinancialSettingsResponseDto
     */
    ico: string;
    /**
     * 
     * @type {string}
     * @memberof UserFinancialSettingsResponseDto
     */
    dic: string;
    /**
     * 
     * @type {string}
     * @memberof UserFinancialSettingsResponseDto
     */
    companyName: string;
    /**
     * 
     * @type {boolean}
     * @memberof UserFinancialSettingsResponseDto
     */
    isCompany: boolean;
    /**
     * 
     * @type {string}
     * @memberof UserFinancialSettingsResponseDto
     */
    iban: string;
    /**
     * 
     * @type {string}
     * @memberof UserFinancialSettingsResponseDto
     */
    swift: string;
    /**
     * 
     * @type {UserAddressResponseDto}
     * @memberof UserFinancialSettingsResponseDto
     */
    address?: UserAddressResponseDto;
    /**
     * 
     * @type {UserAddressResponseDto}
     * @memberof UserFinancialSettingsResponseDto
     */
    mailingAddress?: UserAddressResponseDto;
}